renamed coordinates model file and moved api fetch error handling to a better place
