+++ SAKPaaS/src/app/core/models/position-coordinates.model.ts
